moved CustomLogoutView out of views.py to user.py
