+++ django/cantusdb_project/main_app/urls.py
@@ -4,7 +4,7 @@ from main_app.views import views
 from main_app.views.source import SourceCreateView, SourceEditView
 from main_app.views.chant import ChantEditVolpianoView
 from django.contrib.auth import views as auth_views
-from main_app.views.user import UserDetailView, UserSourceListView
+from main_app.views.user import UserDetailView, UserSourceListView, CustomLogoutView
 
 urlpatterns = [
     # static pages
@@ -12,7 +12,7 @@ urlpatterns = [
     path("contact/", views.contact_us, name="contact"),
     # login/logout/user
     path('login/', auth_views.LoginView.as_view(redirect_authenticated_user=True), name="login"),
-    path('logout/', views.CustomLogoutView.as_view(), name="logout"),
+    path('logout/', CustomLogoutView.as_view(), name="logout"),
     path("my-sources/", UserSourceListView.as_view(), name="my-sources"),
     path("users/<int:user_id>", UserDetailView.as_view(), name="user-detail"),
